complete model end

+++ database.py
@@ -1,4 +1,5 @@
 import os
+import argparse
 
 from flask import Flask, request, send_from_directory
 from flask_sqlalchemy import SQLAlchemy
@@ -9,6 +10,8 @@ from sqlalchemy.orm import relationship
 # from flask_sqlalchemy import SQLAlchemy
 from datetime import datetime, timedelta
 
+from constants import *
+
 # __all__ = ['Dataset']
 app = Flask(__name__)
 app.config.from_object(os.environ['APP_SETTINGS'])
@@ -17,26 +20,23 @@ db = SQLAlchemy(app)
 class Dataset(db.Model):
     __tablename__ = "Dataset"
     # db.Column(primary_key, autoincrement, default, nullable, unique, onupdate, name)
-    name = db.Column(db.String(32), primary_key=True, nullable=False)
-    path = db.Column(db.String(128), nullable=False)
-    audio_dir = db.Column(db.String(64), nullable=False)
-    faces_dir = db.Column(db.String(64), nullable=False)
-    label_path = db.Column(db.String(64), nullable=False)
-    language = db.Column(db.String(16), nullable=False, default="en")
-    label_type = db.Column(db.String(16), nullable=False, default="classification")
+    # dataset_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
+    dataset_name = db.Column(db.String(32), primary_key=True, nullable=False)
+    dataset_path = db.Column(db.String(128), nullable=False)
+    language = db.Column(db.String(16), nullable=False, default="EN")
+    label_path = db.Column(db.String(32), nullable=False)
     text_format = db.Column(db.String(8), nullable=False, default='txt')
     audio_format = db.Column(db.String(8), nullable=False, default='wav')
     video_format = db.Column(db.String(8), nullable=False, default='mp4')
+    raw_video_dir = db.Column(db.String(64), nullable=False)
+    audio_dir = db.Column(db.String(64), nullable=False)
+    faces_dir = db.Column(db.String(64), nullable=False)
     has_feature = db.Column(db.Boolean, nullable=False, default=False)
     is_locked = db.Column(db.Boolean, nullable=False, default=False)
     description = db.Column(db.String(128))
 
-    def get_id(self):
-        return str(self.id)
-
     def __repr__(self):
         return str(self.__dict__)
-
 
 class Dsample(db.Model):
     __tablename__ = "Dsample"
@@ -45,10 +45,12 @@ class Dsample(db.Model):
     video_id = db.Column(db.String(32), nullable=False)
     clip_id = db.Column(db.String(32), nullable=False)
     video_path = db.Column(db.String(128), nullable=False)
-    text = db.Column(db.String(256), nullable=False)
-    sample_mode = db.Column(db.String(8), nullable=False) # train / valid / test
+    text = db.Column(db.String(MAX_TEXT_LEN), nullable=False)
+    # 0 -- train, 1 -- valid, 2 -- test
+    data_mode = db.Column(db.String(8), nullable=False)
     label_value = db.Column(db.Float)
-    label_by = db.Column(db.Integer, nullable=False) # -1 - unlabelled, 0 - human, 1 - machine, 2 - middle, 3 - hard
+    # -1 - unlabelled, 0 - human, 1 - machine, 2 - middle, 3 - hard
+    label_by = db.Column(db.Integer, nullable=False) 
 
     def __repr__(self):
         return str(self.__dict__)
@@ -56,9 +58,10 @@ class Dsample(db.Model):
 class Dfeature(db.Model):
     __tablename__ = "Dfeature"
 
-    dataset_name = db.Column(db.String(32), primary_key=True, nullable=False)
+    feature_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
+    dataset_name = db.Column(db.String(32), nullable=False)
     feature_path = db.Column(db.String(128), nullable=False)
-    input_lens = db.Column(db.String(32), nullable=False)
+    seq_lens = db.Column(db.String(32), nullable=False)
     feature_dims = db.Column(db.String(32), nullable=False)
     description = db.Column(db.String(128))
 
@@ -67,14 +70,12 @@ class Dfeature(db.Model):
 
 class Model(db.Model):
     __tablename__ = "Model"
-    # db.Column(primary_key, autoincrement, default, nullable, unique, onupdate, name)
-    name = db.Column(db.String(32), primary_key=True, nullable=False)
-    common_params = db.Column(db.String(256), nullable=False, default="{}")
-    specific_params = db.Column(db.String(256), nullable=False, default="{}")
+    # model_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
+    model_name = db.Column(db.String(32), primary_key=True, nullable=False)
+    args = db.Column(db.String(MAX_ARGS_LEN), nullable=False, default="{}")
+    paper_name = db.Column(db.String(128))
+    paper_url = db.Column(db.String(128))
     description = db.Column(db.String(128))
-
-    def get_id(self):
-        return str(self.id)
 
     def __repr__(self):
         return str(self.__dict__)
@@ -82,13 +83,15 @@ class Model(db.Model):
 class Result(db.Model):
     __tablename__ = "Result"
     result_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    dataset_name = db.Column(db.String(32), ForeignKey('Dataset.name'), nullable=False)
-    model_name = db.Column(db.String(32), ForeignKey('Model.name'), nullable=False)
-    is_tuning = db.Column(db.Boolean, nullable=False, default=False)
+    dataset_name = db.Column(db.String(32), nullable=False)
+    model_name = db.Column(db.String(32), nullable=False)
+    data_mode = db.Column(db.String(8), nullable=False)
+    # tune, normal
+    is_tuning = db.Column(db.String(8), nullable=False)
     created_at = db.Column(db.Date, default=datetime.utcnow() + timedelta(hours=8))
-    params = db.Column(db.String(1024), nullable=False, default="{}")
-    key_eval_name = db.Column(db.String(16), nullable=False)
-    key_eval_value = db.Column(db.Float, nullable=False)
+    args = db.Column(db.String(MAX_ARGS_LEN), nullable=False, default="{}")
+    save_model_path = db.Column(db.String(128))
+    # description = db.Column(db.String(128))
 
     def get_id(self):
         return str(self.result_id)
@@ -96,32 +99,17 @@ class Result(db.Model):
     def __repr__(self):
         return str(self.__dict__)
 
-class CLResult(db.Model):
-    __tablename__ = "CLResult"
-    result_id = db.Column(db.Integer, primary_key=True, nullable=True)
-    epoch_num = db.Column(db.Integer, nullable=False)
-    loss_value = db.Column(db.Float)
-    accuracy = db.Column(db.Float)
-    macro_f1 = db.Column(db.Float)
-
-    def __repr__(self):
-        return str(self.__dict__)
-
-class RgResult(db.Model):
-    __tablename__ = "RgResult"
-    result_id = db.Column(db.Integer, primary_key=True, nullable=True)
-    epoch_num = db.Column(db.Integer, nullable=False)
-    loss_value = db.Column(db.Float)
-    mse = db.Column(db.Float)
-    corr = db.Column(db.Float)
-
-    def __repr__(self):
-        return str(self.__dict__)
-
-class ModelSave(db.Model):
-    __tablename__ = "ModelSave"
-    result_id = db.Column(db.Integer, primary_key=True, nullable=True)
-    model_path = db.Column(db.String(128)) # model_name: 模型名字__数据集名字__KeyEval结果值__日期
+class EResult(db.Model):
+    # results for each epoch
+    # epoch_num == -1 means the final results
+    __tablename__ = "EResult"
+    result_id = db.Column(db.Integer, primary_key=True, nullable=False)
+    epoch_num = db.Column(db.Integer, primary_key=True, nullable=False)
+    loss_value = db.Column(db.String(32))
+    accuracy = db.Column(db.String(32))
+    f1 = db.Column(db.String(32))
+    mae = db.Column(db.String(32))
+    corr = db.Column(db.String(32))
 
     def __repr__(self):
         return str(self.__dict__)
@@ -129,19 +117,28 @@ class ModelSave(db.Model):
 class Task(db.Model):
     __tablename__ = "Task"
     task_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    dataset_name = db.Column(db.String(32), ForeignKey('Dataset.name'), nullable=False)
-    model_name = db.Column(db.String(32), ForeignKey('Model.name'), nullable=False)
-    # 0 - 主动学习预训练，1 - 特征预处理，2 - 模型训练，3 - 模型调参，4 - 模型测试
+    dataset_name = db.Column(db.String(32), nullable=False)
+    model_name = db.Column(db.String(32), nullable=False)
+    # 0 - 机器标注，1 - 特征预处理，2 - 模型训练，3 - 模型调参，4 - 模型测试
     task_type = db.Column(db.Integer, nullable=False)
-    pid = db.Column(db.Integer, nullable=False)
-    state = db.Column(db.Integer, nullable=False) # 0 -- 运行中，1 -- 已完成，2 -- 出错
+    task_pid = db.Column(db.Integer, nullable=False)
+    # 0 -- 运行中，1 -- 已完成，2 -- 运行出错
+    state = db.Column(db.Integer, nullable=False) 
     start_time = db.Column(db.Date, default=datetime.utcnow() + timedelta(hours=8))
     end_time = db.Column(db.Date, default=datetime.utcnow() + timedelta(hours=8))
-    error_info = db.Column(db.String(128))
+    message = db.Column(db.String(32))
 
     def __repr__(self):
         return str(self.__dict__)
 
+def parse_args():
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--mode', type=str)
+    return parser.parse_args()
+
 if __name__ == '__main__':
-    db.drop_all()
-    db.create_all()
+    arg = parse_args()
+    if arg.mode == "all" or arg.mode == "drop":
+        db.drop_all()
+    if arg.mode == "all" or arg.mode == "create":
+        db.create_all()
